Initial higher-level ei event API, used in receive example
Still needs some more error handling, and handling of device/interface
removal.

This and the handshake can be combined into a very straightforward async
API. A wrapper for that is still needed. The lower level API here could
also be used without async.

An eis event API is also needed for use in the compositor.